modified code for icec_asmr2 ioda converter
