+++ utils/obsproc/applications/gdas_obsprovider2ioda.h
@@ -6,7 +6,7 @@
 #include "oops/mpi/mpi.h"
 #include "oops/runs/Application.h"
 
-#include "../Amsr2Ioda.h"
+#include "../IcecAmsr2Ioda.h"
 #include "../Rads2Ioda.h"
 
 namespace gdasapp {
@@ -21,11 +21,11 @@ namespace gdasapp {
       std::string provider;
       fullConfig.get("provider", provider);
 
-      if (provider == "AMSR2") {
-        Amsr2Ioda conv2ioda(fullConfig);
-        conv2ioda.writeToIoda();
-      } else if (provider == "RADS") {
+      if (provider == "RADS") {
         Rads2Ioda conv2ioda(fullConfig);
+        conv2ioda.writeToIoda();
+      } else if (provider == "AMSR2") {
+        IcecAmsr2Ioda conv2ioda(fullConfig);
         conv2ioda.writeToIoda();
       } else if (provider == "GHRSST") {
         oops::Log::info() << "Comming soon!" << std::endl;
